backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
openai==1.84.0
pyPDF2==3.0.1
docx2txt==0.9
requests==2.32.3
geopy==2.4.1
python-dotenv==1.1.0
asyncio==3.4.3
aiofiles==24.1.0
apscheduler==3.11.0
pydantic==2.11.5
asyncio==3.4.3
statistics==1.0.3.5
gspread==6.2.1
pandas==2.3.0
pytest==8.4.1
pytest-asyncio==1.0.0
httpx==0.28.1